refactor(benchmark): isolate Gin bundle validation
